backend/gallery/views.py
from rest_framework import viewsets
from rest_framework.permissions import AllowAny
from .models import Image
from .serializers import ImageSerializer


class ImageViewSet(viewsets.ModelViewSet):
    queryset = Image.objects.all()
    serializer_class = ImageSerializer
