backend/requirements.txt
asgiref==3.8.1
boto3==1.35.31
botocore==1.35.31
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==0.61.1
django-cors-headers==4.4.0
django-sequences==3.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
load-dotenv==0.1.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
